% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_images.R
\name{classify_images}
\alias{classify_images}
\title{Classify images to database}
\usage{
classify_images(db = NULL, database_directory, photo_directory = NULL)
}
\arguments{
\item{db}{NULL or a suitably formatted database. If NULL, it will fetch the most recently time-stamped file in the database directory.}

\item{database_directory}{Folder containing databse file}

\item{photo_directory}{NULL or directory where photos are stored. Not currently used for anything.}
}
\description{
Prompt user with options for bathymetric image classification
}
